Crate cfx_packing_pool

Structs

MockTransaction

A minimal implementation of the PackingPoolTransaction trait for testing purposes.

PackingBatch

A batch of transactions that have the same sender and continuous nonces.

PackingPool

A PackingPool implementing random packing algorithm and supporting packing a series of transactions with the same nonce.

PackingPoolConfig

Configuration settings for a PackingBatch.

TxSampler

An iterator for sampling transactions from a packing pool.

Enums

SampleTag

Enum representing the phase in which a transaction was selected during the sampling process. See TxSampler for more details.

Traits

PackingPoolTransaction

Trait representing a transaction processed by the PackingPool.